requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
coverage==7.13.5
cryptography==46.0.6
ecdsa==0.19.2
fastapi==0.135.3
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
ply==3.11
pyasn1==0.6.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pysmi-lextudio==1.4.3
pysnmp-lextudio==6.3.0
pysnmpcrypto==0.0.4
pytest==9.0.2
pytest-cov==4.1.0
python-dotenv==1.2.2
python-jose==3.5.0
PyYAML==6.0.3
requests==2.33.1
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.43.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
